% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dau_proj_template.R
\name{dau_proj_template}
\alias{dau_proj_template}
\title{Create standard DAU template for RStudio project}
\usage{
dau_proj_template(
  path,
  init_renv = TRUE,
  include_structure_for_pkg,
  create_adhoc_proj,
  include_github_gitignore,
  ...
)
}
\arguments{
\item{path}{path of the new project}

\item{init_renv}{boolean of whether to initiate renv in the project (set to true)}

\item{include_structure_for_pkg}{additional structure for package development}

\item{create_adhoc_proj}{should the folder structure be for an ad-hoc project}

\item{include_github_gitignore}{should a strict gitingore file for GitHub be created}

\item{...}{additional parameters, currently not used}
}
\value{
no return values, just the folders and README are created
}
\description{
Create standard DAU template for RStudio project
}
