% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dau_proj_template.R
\name{standard_analysis}
\alias{standard_analysis}
\title{Create standard analysis template for RStudio project}
\usage{
standard_analysis(
  path,
  include_structure_for_pkg,
  include_github_gitignore,
  ...
)
}
\arguments{
\item{path}{path of the new project}

\item{include_structure_for_pkg}{additional Analysis for publication folder}

\item{include_github_gitignore}{should a gitingore file be created}

\item{...}{additional parameters, currently not used}
}
\value{
no return values, just the folders and README are created
}
\description{
Create standard analysis template for RStudio project
}
