% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dau_proj_template.R
\name{dau_proj_template}
\alias{dau_proj_template}
\title{Create standard analysis template for RStudio project}
\usage{
dau_proj_template(
  path,
  include_structure_for_pkg,
  include_github_gitignore,
  ...
)
}
\arguments{
\item{path}{path of the new project}

\item{include_structure_for_pkg}{additional structure for package development}

\item{include_github_gitignore}{should a strict gitingore file for GitHub be created}

\item{...}{additional parameters, currently not used}
}
\value{
no return values, just the folders and README are created
}
\description{
Create standard analysis template for RStudio project
}
